+++ lib/chef/platform/provider_priority_map.rb
@@ -34,9 +34,10 @@ class Chef
         ], platform_family: "gentoo"
 
         priority :service, [
-          # on debian-ish system if an upstart script exists that always wins
-          Chef::Provider::Service::Upstart,
+          # we can determine what systemd supports accurately
           Chef::Provider::Service::Systemd,
+          # on debian-ish system if an upstart script exists that must win over sysv types
+          Chef::Provider::Service::Upstart,
           Chef::Provider::Service::Insserv,
           Chef::Provider::Service::Debian,
           Chef::Provider::Service::Invokercd,