+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Chef Contributions
+* Update Chef to use RSpec 3.
+
+
 ## 12.0.0
 
 * [**Jesse Hu**](https://github.com/jessehu):
@@ -237,7 +243,7 @@
 * The Windows env provider will delete elements even if they are only in ENV (and not in the registry)
 * Allow events to be logged to Windows Event Log
 * Fixed bug in env resource where a value containing the delimiter could never correctly match the existing values
-* Update Chef to use RSpec 3.
+* More intelligent service check for systemd on Ubuntu 14.10.
 
 ## 11.16.4